- second step : the skin is now available (problem to solve : the SharedDic doesn't seem t be able to find CK.Keyboard.KeyboardCollection when CK.Keyboard.dll is not directly in the output folder, where it shouldn't be)

+++ Plugins/Accessibility/SimpleSkin/SkinWindow.xaml.cs
@@ -56,7 +56,7 @@ namespace SimpleSkin
         {
             _interopHelper = new WindowInteropHelper( this );
 
-            Win.Functions.SetWindowLong( _interopHelper.Handle, Win.WindowLongIndex.GWL_EXSTYLE, (long)Win.WS_EX.NOACTIVATE );
+            Win.Functions.SetWindowLong( _interopHelper.Handle, Win.WindowLongIndex.GWL_EXSTYLE, (uint)Win.WS_EX.NOACTIVATE );
 
             HwndSource mainWindowSrc = HwndSource.FromHwnd( _interopHelper.Handle );
